+++ src/store/index.js
@@ -89,9 +89,11 @@ const migrations = [
         },
     },
     {
-        version: 3,
+        version: 4,
         up: (state) => {
-            const orgFavorites = state.orgFavorites.filter((v) => v !== "Virtual Witch Phenomenon");
+            const orgFavorites = state.orgFavorites.filter(
+                (v) => v !== "Virtual Witch Phenomenon" && v !== "Hanayori Joshiryo",
+            );
             return {
                 ...state,
                 orgFavorites,